requirements.txt
streamlit
python-dotenv
google-generativeai
psycopg2-binary
